requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
websockets==11.0.3
python-dotenv==1.0.0